+++ manage/admin/manage_ins_diagnosis.php
@@ -1,7 +1,7 @@
 <? 
 include "includes/top.htm";
 
-if($_REQUEST["delid"] != "")
+if($_REQUEST["delid"] != "" && $_SESSION['admin_access']==1)
 {
 	$del_sql = "delete from dental_ins_diagnosis where ins_diagnosisid='".$_REQUEST["delid"]."'";
 	mysql_query($del_sql);
